requirements.txt
numpy
matplotlib
scipy
torch
torchvision
hydra-core
omegaconf
phiflow
h5py
goppy
tensorboard
tqdm
scikit-image
pandas
cvxpy
scikit-fmm
